requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
